requirements.txt
allure-pytest==2.8.40
allure-python-commons == 2.8.40
mysql-connector-python==8.0.24
openpyxl==3.0.7
pytest==6.2.3
pytest-forked==1.3.0
pytest-html== 3.1.1
pytest-metadata==1.11.0
pytest-rerunfailures==9.1.1
pytest-xdist==2.2.1
requests ==2.25.1
selenium==3.141.0
setuptools==49.2.1
six ==1.15.0
urllib3==1.26.4
webdriver-manager
pandas
numpy
